Cherry pick library renaming commit
