+++ MANIFEST.in
@@ -1,2 +1,1 @@
-recursive-include src/robothub_oak *.py
-recursive-include src/robothub_oak/data_processors *.py
+recursive-include src/robothub *.py
