+++ README.md
@@ -1,4 +1,4 @@
-# RobotHub-OAK
+# RobotHub
 
 This repository contains the code for the integration of [DepthAI SDK](https://github.com/luxonis/depthai) into
 RobotHub.
@@ -14,7 +14,7 @@ Please visit [RobotHub documentation page](https://hub-docs.luxonis.com/).
 To install this package, run the following command in your terminal window
 
 ```
-$ python3 -m pip install robothub-oak
+$ python3 -m pip install robothub
 ```
 
 ## Examples
